move test setup/helpers to a separate module
